CLIP_FAISS/backend/requirements.txt
# ========================================
# 企业产品以图搜系统 - Python依赖清单
# ========================================

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 深度学习框架
torch>=2.0.0
transformers>=4.30.0

# 图像处理
pillow==10.1.0

# 向量检索
faiss-cpu>=1.7.0
# 如需GPU加速，请使用：faiss-gpu==1.7.4

# 数据库
pymysql>=1.0.0
sqlalchemy>=2.0.0

# 环境变量管理
python-dotenv>=1.0.0

# 日志和监控
loguru>=0.7.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 测试工具
pytest==7.4.3
httpx==0.25.2

# 代码质量
black==23.11.0
flake8==6.1.0
